refactor : PostPage.js로 postlist불러오도록 변경
