+++ src/components/PostPage.js
@@ -1,0 +1,31 @@
+import PostList from "./PostList.js"
+import { request } from "../api/api.js"
+
+
+export default function PostPage({
+    $target
+}) {
+    const $postPage = document.createElement('div')
+    $postPage.className  = "post-page"
+    
+    
+    
+    const postList = new PostList({
+        $target,
+        initialState :[]
+    })
+
+    
+const fetchPosts = async() => {
+    const posts = await request('/')
+    postList.setState(posts)
+  }
+  
+  fetchPosts()
+    
+
+    this.render = () => {
+        $target.appendChild($postPage)
+
+    }
+}
